+++ foc_core/include/L2_Core/Control/foc_ctrl_transition.h
@@ -5,7 +5,6 @@
 
 #include "L2_Core/foc_ctrl_types.h"
 
-void FOC_Transition_OnSourceSwitch(foc_motor_t *motor, uint8_t new_source, uint8_t old_source);
 void FOC_Transition_OnModeSwitch(foc_motor_t *motor, uint8_t new_mode, uint8_t old_mode);
 
 #endif /* FOC_CTRL_TRANSITION_H */